Moved all subtests in the same file
Just moved functions around, no code change.
