+++ internal/integrationtest/compile/compile_part_1_test.go
@@ -26,6 +26,7 @@ import (
 	"github.com/arduino/arduino-cli/internal/integrationtest"
 	"github.com/arduino/go-paths-helper"
 	"github.com/stretchr/testify/require"
+	"go.bug.st/testifyjson/requirejson"
 )
 
 func TestCompile(t *testing.T) {
@@ -229,60 +230,6 @@ func compileBlacklistedSketchname(t *testing.T, env *integrationtest.Environment
 	require.NoError(t, err)
 }
 
-func TestCompileWithoutPrecompiledLibraries(t *testing.T) {
-	env, cli := integrationtest.CreateArduinoCLIWithEnvironment(t)
-	defer env.CleanUp()
-
-	// Init the environment explicitly
-	url := "https://adafruit.github.io/arduino-board-index/package_adafruit_index.json"
-	_, _, err := cli.Run("core", "update-index", "--additional-urls="+url)
-	require.NoError(t, err)
-	_, _, err = cli.Run("core", "install", "arduino:mbed@1.3.1", "--additional-urls="+url)
-	require.NoError(t, err)
-
-	// // Precompiled version of Arduino_TensorflowLite
-	//	_, _, err = cli.Run("lib", "install", "Arduino_LSM9DS1")
-	//	require.NoError(t, err)
-	//	_, _, err = cli.Run("lib", "install", "Arduino_TensorflowLite@2.1.1-ALPHA-precompiled")
-	//	require.NoError(t, err)
-
-	//	sketchPath := cli.SketchbookDir().Join("libraries", "Arduino_TensorFlowLite", "examples", "hello_world")
-	//	_, _, err = cli.Run("compile", "-b", "arduino:mbed:nano33ble", sketchPath.String())
-	//	require.NoError(t, err)
-
-	_, _, err = cli.Run("core", "install", "arduino:samd@1.8.7", "--additional-urls="+url)
-	require.NoError(t, err)
-	//	_, _, err = cli.Run("core", "install", "adafruit:samd@1.6.4", "--additional-urls="+url)
-	//	require.NoError(t, err)
-	//	// should work on adafruit too after https://github.com/arduino/arduino-cli/pull/1134
-	//	_, _, err = cli.Run("compile", "-b", "adafruit:samd:adafruit_feather_m4", sketchPath.String())
-	//	require.NoError(t, err)
-
-	//	// Non-precompiled version of Arduino_TensorflowLite
-	//	_, _, err = cli.Run("lib", "install", "Arduino_TensorflowLite@2.1.0-ALPHA")
-	//	require.NoError(t, err)
-	//	_, _, err = cli.Run("compile", "-b", "arduino:mbed:nano33ble", sketchPath.String())
-	//	require.NoError(t, err)
-	//	_, _, err = cli.Run("compile", "-b", "adafruit:samd:adafruit_feather_m4", sketchPath.String())
-	//	require.NoError(t, err)
-
-	// Bosch sensor library
-	_, _, err = cli.Run("lib", "install", "BSEC Software Library@1.5.1474")
-	require.NoError(t, err)
-	sketchPath := cli.SketchbookDir().Join("libraries", "BSEC_Software_Library", "examples", "basic")
-	_, _, err = cli.Run("compile", "-b", "arduino:samd:mkr1000", sketchPath.String())
-	require.NoError(t, err)
-	_, _, err = cli.Run("compile", "-b", "arduino:mbed:nano33ble", sketchPath.String())
-	require.NoError(t, err)
-
-	// USBBlaster library
-	_, _, err = cli.Run("lib", "install", "USBBlaster@1.0.0")
-	require.NoError(t, err)
-	sketchPath = cli.SketchbookDir().Join("libraries", "USBBlaster", "examples", "USB_Blaster")
-	_, _, err = cli.Run("compile", "-b", "arduino:samd:mkrvidor4000", sketchPath.String())
-	require.NoError(t, err)
-}
-
 func compileWithBuildPropertiesFlag(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
 	{
 		sketchName := "sketch_with_single_string_define"
@@ -394,4 +341,188 @@ func compileWithMultipleBuildPropertyFlags(t *testing.T, env *integrationtest.En
 	require.NoError(t, err)
 	require.Contains(t, string(stdout), "-DPIN=2")
 	require.Contains(t, string(stdout), "-DSSID=\\\"hello world\\\"")
+}
+
+func compileWithOutputDirFlag(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithOutputDir"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	stdout, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+	require.Contains(t, string(stdout), "Sketch created in: "+sketchPath.String())
+
+	// Test the --output-dir flag with absolute path
+	outputDir := cli.SketchbookDir().Join("test_dir", "output_dir")
+	_, _, err = cli.Run("compile", "-b", fqbn, sketchPath.String(), "--output-dir", outputDir.String())
+	require.NoError(t, err)
+
+	// Verifies expected binaries have been built
+	md5 := md5.Sum(([]byte(sketchPath.String())))
+	sketchPathMd5 := strings.ToUpper(hex.EncodeToString(md5[:]))
+	require.NotEmpty(t, sketchPathMd5)
+	buildDir := paths.TempDir().Join("arduino-sketch-" + sketchPathMd5)
+	require.FileExists(t, buildDir.Join(sketchName+".ino.eep").String())
+	require.FileExists(t, buildDir.Join(sketchName+".ino.elf").String())
+	require.FileExists(t, buildDir.Join(sketchName+".ino.hex").String())
+	require.FileExists(t, buildDir.Join(sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, buildDir.Join(sketchName+".ino.with_bootloader.hex").String())
+
+	// Verifies binaries are exported when --output-dir flag is specified
+	require.DirExists(t, outputDir.String())
+	require.FileExists(t, outputDir.Join(sketchName+".ino.eep").String())
+	require.FileExists(t, outputDir.Join(sketchName+".ino.elf").String())
+	require.FileExists(t, outputDir.Join(sketchName+".ino.hex").String())
+	require.FileExists(t, outputDir.Join(sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, outputDir.Join(sketchName+".ino.with_bootloader.hex").String())
+}
+
+func compileWithExportBinariesFlag(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithExportBinariesFlag"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	_, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+
+	// Test the --output-dir flag with absolute path
+	_, _, err = cli.Run("compile", "-b", fqbn, sketchPath.String(), "--export-binaries")
+	require.NoError(t, err)
+	require.DirExists(t, sketchPath.Join("build").String())
+
+	// Verifies binaries are exported when --export-binaries flag is set
+	fqbn = strings.ReplaceAll(fqbn, ":", ".")
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.eep").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.elf").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.hex").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.hex").String())
+}
+
+func compileWithCustomBuildPath(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithBuildPath"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	_, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+
+	// Test the --build-path flag with absolute path
+	buildPath := cli.DataDir().Join("test_dir", "build_dir")
+	_, _, err = cli.Run("compile", "-b", fqbn, sketchPath.String(), "--build-path", buildPath.String())
+	require.NoError(t, err)
+
+	// Verifies expected binaries have been built to build_path
+	require.DirExists(t, buildPath.String())
+	require.FileExists(t, buildPath.Join(sketchName+".ino.eep").String())
+	require.FileExists(t, buildPath.Join(sketchName+".ino.elf").String())
+	require.FileExists(t, buildPath.Join(sketchName+".ino.hex").String())
+	require.FileExists(t, buildPath.Join(sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, buildPath.Join(sketchName+".ino.with_bootloader.hex").String())
+
+	// Verifies there are no binaries in temp directory
+	md5 := md5.Sum(([]byte(sketchPath.String())))
+	sketchPathMd5 := strings.ToUpper(hex.EncodeToString(md5[:]))
+	require.NotEmpty(t, sketchPathMd5)
+	buildDir := paths.TempDir().Join("arduino-sketch-" + sketchPathMd5)
+	require.NoFileExists(t, buildDir.Join(sketchName+".ino.eep").String())
+	require.NoFileExists(t, buildDir.Join(sketchName+".ino.elf").String())
+	require.NoFileExists(t, buildDir.Join(sketchName+".ino.hex").String())
+	require.NoFileExists(t, buildDir.Join(sketchName+".ino.with_bootloader.bin").String())
+	require.NoFileExists(t, buildDir.Join(sketchName+".ino.with_bootloader.hex").String())
+}
+
+func compileWithExportBinariesEnvVar(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithExportBinariesEnvVar"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	_, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+
+	envVar := cli.GetDefaultEnv()
+	envVar["ARDUINO_SKETCH_ALWAYS_EXPORT_BINARIES"] = "true"
+
+	// Test compilation with export binaries env var set
+	_, _, err = cli.RunWithCustomEnv(envVar, "compile", "-b", fqbn, sketchPath.String())
+	require.NoError(t, err)
+	require.DirExists(t, sketchPath.Join("build").String())
+
+	// Verifies binaries are exported when export binaries env var is set
+	fqbn = strings.ReplaceAll(fqbn, ":", ".")
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.eep").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.elf").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.hex").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.hex").String())
+}
+
+func compileWithExportBinariesConfig(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithExportBinariesEnvVar"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	_, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+
+	// Create settings with export binaries set to true
+	envVar := cli.GetDefaultEnv()
+	envVar["ARDUINO_SKETCH_ALWAYS_EXPORT_BINARIES"] = "true"
+	_, _, err = cli.RunWithCustomEnv(envVar, "config", "init", "--dest-dir", ".")
+	require.NoError(t, err)
+	defer cli.WorkingDir().Join("arduino-cli.yaml").Remove()
+
+	// Test if arduino-cli config file written in the previous run has the `always_export_binaries` flag set.
+	stdout, _, err := cli.Run("config", "dump", "--format", "json")
+	require.NoError(t, err)
+	requirejson.Contains(t, stdout, `
+		{
+			"sketch": {
+			"always_export_binaries": "true"
+	  		}
+		}`)
+
+	// Test compilation with export binaries env var set
+	_, _, err = cli.Run("compile", "-b", fqbn, sketchPath.String())
+	require.NoError(t, err)
+	require.DirExists(t, sketchPath.Join("build").String())
+
+	// Verifies binaries are exported when export binaries env var is set
+	fqbn = strings.ReplaceAll(fqbn, ":", ".")
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.eep").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.elf").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.hex").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.bin").String())
+	require.FileExists(t, sketchPath.Join("build", fqbn, sketchName+".ino.with_bootloader.hex").String())
+}
+
+func compileWithInvalidUrl(t *testing.T, env *integrationtest.Environment, cli *integrationtest.ArduinoCLI) {
+	sketchName := "CompileWithInvalidURL"
+	sketchPath := cli.SketchbookDir().Join(sketchName)
+	defer sketchPath.RemoveAll()
+	fqbn := "arduino:avr:uno"
+
+	// Create a test sketch
+	_, _, err := cli.Run("sketch", "new", sketchPath.String())
+	require.NoError(t, err)
+
+	_, _, err = cli.Run("config", "init", "--dest-dir", ".", "--additional-urls", "https://example.com/package_example_index.json")
+	require.NoError(t, err)
+	defer cli.WorkingDir().Join("arduino-cli.yaml").Remove()
+
+	_, stderr, err := cli.Run("compile", "-b", fqbn, sketchPath.String())
+	require.NoError(t, err)
+	require.Contains(t, string(stderr), "Error initializing instance: Loading index file: loading json index file")
+	expectedIndexfile := cli.DataDir().Join("package_example_index.json")
+	require.Contains(t, string(stderr), "loading json index file "+expectedIndexfile.String()+": open "+expectedIndexfile.String()+":")
 }
